Change the way errors are handled by Rest.
When the http request has completed (ie we receive something from the server),
we do not treat this as an error, regardless of the code sent by the server.

The responsibility of testing for the return code is on the client code. This is
logical, especially when code updating contacts can perfectly expect a 412 from
Google, or code getting images must expect to have 404 for example. These are
not exceptional situations. Rest object should only care if it manages to send a
request and receive a response.