Correctly remember etags and ignore unchanged contacts.
We now store the new entry we get from google when we query the list of changes
or when we update a contact.
Then we can compare etags when querying the google contacts api, and ignore
every contact that has an unchanged etag. This is the case when we updated
this contact during previous sync: the change is after the lastSyncDate,
therefore this modification we made will appear in new sync feed, and we should
ignore it.