Send correct headers to google.
Previously, the headers were not copied correctly, resulting in etags leaking
from one request to another.

Also, this adds a suitable Content-Type to the headers. It seems gecko was able
to put the right one most of the time, but it seems not to be the case every
time for POST request we need for adding a contact.